Add files via upload
This is an attempt to represent the NALBB grid based on the Voronoï-Thiessen polygon algorithm. Since the atlas does not cover the entire villages of Lower-Brittany, this algorithm can extend each survey point as polygons to the next surrounding points. Changes will have to be made in order to correct the overlapping of bays and rivers.